move rmtree() to file_util.py
