+++ mock/py/mockbuild/backend.py
@@ -113,30 +113,35 @@ class Commands(object):
                     if scrub == 'all':
                         self.buildroot.root_log.info("scrubbing everything for %s", self.config_name)
                         self.buildroot.delete()
-                        util.rmtree(self.buildroot.cachedir, selinux=self.buildroot.selinux)
+                        file_util.rmtree(self.buildroot.cachedir, selinux=self.buildroot.selinux)
                         if self.bootstrap_buildroot is not None:
                             self.bootstrap_buildroot.delete()
-                            util.rmtree(self.bootstrap_buildroot.cachedir, selinux=self.bootstrap_buildroot.selinux)
+                            file_util.rmtree(self.bootstrap_buildroot.cachedir,
+                                             selinux=self.bootstrap_buildroot.selinux)
                     elif scrub == 'chroot':
                         self.buildroot.root_log.info("scrubbing chroot for %s", self.config_name)
                         self.buildroot.delete()
                     elif scrub == 'cache':
                         self.buildroot.root_log.info("scrubbing cache for %s", self.config_name)
-                        util.rmtree(self.buildroot.cachedir, selinux=self.buildroot.selinux)
+                        file_util.rmtree(self.buildroot.cachedir, selinux=self.buildroot.selinux)
                     elif scrub == 'c-cache':
                         self.buildroot.root_log.info("scrubbing c-cache for %s", self.config_name)
-                        util.rmtree(os.path.join(self.buildroot.cachedir, 'ccache'), selinux=self.buildroot.selinux)
+                        file_util.rmtree(os.path.join(self.buildroot.cachedir, 'ccache'),
+                                         selinux=self.buildroot.selinux)
                     elif scrub == 'root-cache':
                         self.buildroot.root_log.info("scrubbing root-cache for %s", self.config_name)
-                        util.rmtree(os.path.join(self.buildroot.cachedir, 'root_cache'), selinux=self.buildroot.selinux)
+                        file_util.rmtree(os.path.join(self.buildroot.cachedir, 'root_cache'),
+                                         selinux=self.buildroot.selinux)
                     elif scrub in ['yum-cache', 'dnf-cache']:
                         self.buildroot.root_log.info("scrubbing yum-cache and dnf-cache for %s", self.config_name)
-                        util.rmtree(os.path.join(self.buildroot.cachedir, 'yum_cache'), selinux=self.buildroot.selinux)
-                        util.rmtree(os.path.join(self.buildroot.cachedir, 'dnf_cache'), selinux=self.buildroot.selinux)
+                        file_util.rmtree(os.path.join(self.buildroot.cachedir, 'yum_cache'),
+                                         selinux=self.buildroot.selinux)
+                        file_util.rmtree(os.path.join(self.buildroot.cachedir, 'dnf_cache'),
+                                         selinux=self.buildroot.selinux)
                     elif scrub == 'bootstrap' and self.bootstrap_buildroot is not None:
                         self.buildroot.root_log.info("scrubbing bootstrap for %s", self.config_name)
                         self.bootstrap_buildroot.delete()
-                        util.rmtree(self.bootstrap_buildroot.cachedir, selinux=self.bootstrap_buildroot.selinux)
+                        file_util.rmtree(self.bootstrap_buildroot.cachedir, selinux=self.bootstrap_buildroot.selinux)
 
             except IOError as e:
                 getLog().warning("parts of chroot do not exist: %s", e)
@@ -562,7 +567,7 @@ class Commands(object):
                 # Resolve any symlinks
                 sources = os.path.realpath(sources)
                 if os.path.isdir(sources):
-                    util.rmtree(self.buildroot.make_chroot_path(self.buildroot.builddir, "SOURCES"))
+                    file_util.rmtree(self.buildroot.make_chroot_path(self.buildroot.builddir, "SOURCES"))
                     shutil.copytree(sources,
                                     self.buildroot.make_chroot_path(self.buildroot.builddir, "SOURCES"),
                                     symlinks=(not follow_links))
